backend/requirements.txt
flask
flask-cors
flask-socketio
flask-swagger-ui
python-dotenv
simple-websocket

torch
pytorch-lightning
transformers
datasets
sentencepiece
protobuf
sacrebleu
rouge-score
bert-score

tika
apscheduler
beautifulsoup4

fairseq